--- src/lib/legacy/thank-you/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-	import groupPhoto from '$lib/assets/slice-a-thon_group_photo.webp?inline';
-</script>
-
-<div class="my-5 place-items-center">
-	<img class="my-4" src={groupPhoto} alt="group" />
-	<h1 class="text-red-sauce text-center text-3xl">Thank you!</h1>
-</div>
