+++ src/routes/eggnog/+layout.svelte
@@ -1,0 +1,11 @@
+<script>
+	let { children } = $props();
+</script>
+
+<div class="flex w-full items-center justify-center bg-amber-400">
+	<a href="/eggnog/2025">
+		<button> 2025 </button>
+	</a>
+</div>
+
+{@render children()}
