+++ src/routes/eggnog/+page.ts
@@ -1,0 +1,6 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load() {
+    redirect(307, '/eggnog/2025');
+}
+
