+++ src/routes/slice-a-thon/+layout.svelte
@@ -1,0 +1,11 @@
+<script>
+	let { children } = $props();
+</script>
+
+<div class="flex w-full items-center justify-center bg-amber-400">
+	<a href="/slice-a-thon/2025">
+		<button> 2025 </button>
+	</a>
+</div>
+
+{@render children()}
