+++ src/routes/slice-a-thon/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load() {
+    redirect(307, '/slice-a-thon/2025');
+}
