+++ src/routes/taco-ocho/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load() {
+    redirect(307, '/taco-ocho/2025');
+}
